feat: 메인페이지 코드 스플리팅
