+++ src/features/main/components/sections/index.ts
@@ -1,0 +1,3 @@
+export * from './event-section'
+export * from './group-challenge-section'
+export * from './personal-challenge-section'
